fix_md/requirements.txt
certifi==2024.6.2
quickfix-ssl==1.15.1.post4